refactor: Ready가 Started 상태를 상속하도록 변경
